create new @cliqz/adblocker-content package with common utils
