+++ packages/adblocker-circumvention/package.json
@@ -42,6 +42,6 @@
     "typescript": "^3.5.3"
   },
   "dependencies": {
-    "@cliqz/adblocker": "^0.12.1"
+    "@cliqz/adblocker-content": "^0.12.1"
   }
 }
