+++ packages/adblocker-circumvention/tsconfig.json
@@ -7,7 +7,7 @@
   },
   "references": [
     {
-      "path": "../adblocker/tsconfig.json"
+      "path": "../adblocker-content/tsconfig.json"
     }
   ],
   "include": [
