+++ packages/adblocker-content/README.md
@@ -1,0 +1,3 @@
+# `@cliqz/adblocker-content`
+
+> Helpers used exclusively in content-scripts contexts
