+++ packages/adblocker-content/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@cliqz/adblocker-webextension-cosmetics",
+  "name": "@cliqz/adblocker-content",
   "version": "0.12.1",
-  "description": "Enable cosmetics in WebExtension content blocker using Cliqz adblocker",
+  "description": "Cliqz adblocker library (content-scripts helpers)",
   "author": "Cliqz",
   "homepage": "https://github.com/cliqz-oss/adblocker#readme",
   "license": "MPL-2.0",
@@ -25,8 +25,7 @@
     "build": "tsc --build ./tsconfig.json",
     "bundle": "rollup --config ./rollup.config.ts",
     "prebundle": "yarn run build",
-    "prepack": "yarn run bundle",
-    "test": "jest --config ./jest.config.js --coverage ./adblocker.test.ts"
+    "prepack": "yarn run bundle"
   },
   "bugs": {
     "url": "https://github.com/cliqz-oss/adblocker/issues"
@@ -35,16 +34,16 @@
     "@ampproject/rollup-plugin-closure-compiler": "^0.10.0",
     "@types/chrome": "^0.0.88",
     "@types/jest": "^24.0.15",
-    "@types/jsdom": "^12.2.4",
+    "@types/node": "^12.0.3",
     "jest": "^24.8.0",
-    "jsdom": "^15.1.1",
     "rimraf": "^2.6.3",
     "rollup": "^1.16.4",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-sourcemaps": "^0.4.2",
+    "ts-jest": "^24.0.2",
+    "ts-node": "^8.3.0",
+    "tslint": "^5.18.0",
+    "tslint-config-prettier": "^1.18.0",
     "typescript": "^3.5.3"
-  },
-  "dependencies": {
-    "@cliqz/adblocker-content": "^0.12.1"
   }
 }
