+++ packages/adblocker-content/tsconfig.json
@@ -5,10 +5,7 @@
     "declarationDir": "dist/types",
     "composite": true
   },
-  "references": [
-    { "path": "../adblocker-content/tsconfig.json" }
-  ],
-  "files": [
+  "include": [
     "adblocker.ts"
   ]
 }
