+++ packages/adblocker-electron-example/tsconfig.json
@@ -6,7 +6,7 @@
     "composite": true
   },
   "references": [
-    { "path": "../adblocker-electron" }
+    { "path": "../adblocker-electron/tsconfig.json" }
   ],
   "files": [
     "index.ts"
