+++ packages/adblocker-electron/package.json
@@ -36,7 +36,7 @@
   },
   "dependencies": {
     "@cliqz/adblocker": "^0.12.1",
-    "@cliqz/adblocker-webextension-cosmetics": "^0.12.1"
+    "@cliqz/adblocker-content": "^0.12.1"
   },
   "devDependencies": {
     "jest": "^24.8.0",
