+++ packages/adblocker-electron/tsconfig.json
@@ -7,7 +7,7 @@
   },
   "references": [
     { "path": "../adblocker/tsconfig.json" },
-    { "path": "../adblocker-webextension-cosmetics/tsconfig.json" }
+    { "path": "../adblocker-content/tsconfig.json" }
   ],
   "files": [
     "adblocker.ts",
