+++ packages/adblocker-puppeteer-example/tsconfig.json
@@ -6,7 +6,7 @@
     "composite": true
   },
   "references": [
-    { "path": "../adblocker-puppeteer" }
+    { "path": "../adblocker-puppeteer/tsconfig.json" }
   ],
   "files": [
     "index.ts"
