+++ packages/adblocker-puppeteer/package.json
@@ -36,6 +36,7 @@
   },
   "dependencies": {
     "@cliqz/adblocker": "^0.12.1",
+    "@cliqz/adblocker-content": "^0.12.1",
     "@types/puppeteer": "^1.12.4",
     "tldts-experimental": "^5.3.0"
   },
