+++ packages/adblocker-puppeteer/tsconfig.json
@@ -6,7 +6,8 @@
     "composite": true
   },
   "references": [
-    { "path": "../adblocker/tsconfig.json" }
+    { "path": "../adblocker/tsconfig.json" },
+    { "path": "../adblocker-content/tsconfig.json" }
   ],
   "files": [
     "./adblocker.ts"
