+++ packages/adblocker-webextension-example/tsconfig.json
@@ -6,8 +6,8 @@
     "composite": true
   },
   "references": [
-    { "path": "../adblocker-webextension" },
-    { "path": "../adblocker-webextension-cosmetics" }
+    { "path": "../adblocker-webextension/tsconfig.json" },
+    { "path": "../adblocker-webextension-cosmetics/tsconfig.json" }
   ],
   "files": [
     "background.ts",
