+++ packages/adblocker-webextension/package.json
@@ -46,7 +46,7 @@
   },
   "dependencies": {
     "@cliqz/adblocker": "^0.12.1",
-    "@cliqz/adblocker-webextension-cosmetics": "^0.12.1",
+    "@cliqz/adblocker-content": "^0.12.1",
     "tldts-experimental": "^5.3.0"
   }
 }
