+++ packages/tsconfig.project.json
@@ -5,6 +5,9 @@
       "path": "./adblocker/tsconfig.json"
     },
     {
+      "path": "./adblocker-content/tsconfig.json"
+    },
+    {
       "path": "./adblocker-circumvention/tsconfig.json"
     },
     {
